fix: properly inform kernel about partition deletion
When blockdevice is open with new GPT (or in general case, whatever was
the previous partition table), we never know what's the kernel state of
partitions, so we can't clean up via existing partitions on disk, we
need to check kernel state.

This change looks ugly, but it works by simply trying to delete any
partition from the kernel state which is still found in `/sys/block/`
before adding new partitions (which we just got created).

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>