feat: implement "fast" wipe
Full wipe via zeroing out is really slow and might not fit every use
case. Implement "fast" wipe via zeroing out first 1MB of the disk
(should be enough to erase the GPT) and discarding blocks (TRIM for
SSDs) which might zero out the rest of the disk, but it also plays nice
for SSD to mark all blocks as unused.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>