fix: align partition to 1M boundary by default
See e.g. https://www.thomas-krenn.com/en/wiki/Partition_Alignment_detailed_explanation

This is the default used by modern disk partition utilities.

Align partition end as well to the same boundary when the partition is
grown to the maximum size (see
https://github.com/siderolabs/talos/issues/4985).

Fix the problem with `WithSingleResult` which never worked properly (?).

It looks like `probe.All` the way it is designed will never work
properly, as the partition name check is done for the blockdevice as a
whole, while the iteration goes over all partition. This goes unnoticed
because of the way this function is actually used.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>